requirements.txt
temporalio>=1.8.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
databricks-sdk>=0.18.0
databricks-vectorsearch>=0.57
requests>=2.31.0
google-cloud-translate